% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerSurvCoxPH.R
\docType{data}
\name{LearnerSurvCoxPH}
\alias{LearnerSurvCoxPH}
\alias{mlr_learners_surv.coxph}
\title{Cox Proportional Hazard Learner}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{LearnerSurv}.}
\usage{
LearnerSurvCoxPH
}
\description{
A \link{LearnerSurv} for a Cox PH model implemented in \link[survival:coxph]{survival::coxph} in package \CRANpkg{survival}.
}
\keyword{datasets}
