% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerSurvGlmnet.R
\docType{data}
\name{LearnerSurvGlmnet}
\alias{LearnerSurvGlmnet}
\alias{mlr_learners_surv.glmnet}
\title{GLM with Elastic Net Regularization Survival Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{LearnerSurv}.}
\usage{
LearnerSurvGlmnet
}
\description{
Generalized linear models with elastic net regularization.
Calls \code{\link[glmnet:cv.glmnet]{glmnet::cv.glmnet()}} from package \CRANpkg{glmnet}.
}
\examples{
library(mlr3)
task = tgen("simsurv")$generate(200)
learner = lrn("surv.glmnet")
resampling = rsmp("cv", folds = 3)
resample(task, learner, resampling)
}
\references{
Jerome Friedman, Trevor Hastie, Robert Tibshirani (2010).
Regularization Paths for Generalized Linear Models via Coordinate Descent.
Journal of Statistical Software, 33(1), 1-22.
\doi{10.18637/jss.v033.i01}.
}
\keyword{datasets}
