% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerSurvRpart.R
\docType{data}
\name{LearnerSurvRpart}
\alias{LearnerSurvRpart}
\alias{mlr_learners_surv.rpart}
\title{Survival Tree Learner}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{LearnerSurv}.}
\usage{
LearnerSurvRpart
}
\description{
A \link{LearnerSurv} for a regression tree implemented in \link[rpart:rpart]{rpart::rpart} in package \CRANpkg{rpart}.
Parameter \code{xval} has been set to 0 per default.
}
\keyword{datasets}
