% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureSurv.R
\docType{data}
\name{MeasureSurv}
\alias{MeasureSurv}
\title{Survival Measure}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link[mlr3:Measure]{mlr3::Measure}.}
\description{
This measure specializes \link[mlr3:Measure]{mlr3::Measure} for classification problems.
Predefined measures can be found in the \link[mlr3misc:Dictionary]{mlr3misc::Dictionary} \link[mlr3:mlr_measures]{mlr3::mlr_measures}.

The \code{task_type} is set to \code{"classif"}.
}
\section{Construction}{
\preformatted{m = MeasureSurv$new(id, range, minimize, aggregator = NULL, properties = character(), predict_type = "risk",
     task_properties = character(), packages = character())
}

For a description of the arguments, see \link[mlr3:Measure]{mlr3::Measure}.
The \code{task_type} is set to \code{"surv"}.
The currently only possible value for \code{predict_type} is "risk".
}

\section{Fields}{

See \link{Measure}.
}

\section{Methods}{

See \link{Measure}.
}

\seealso{
Example survival measure: \code{\link[=mlr_measures_surv.harrells_c]{surv.harrells_c}}.
}
\concept{Measure}
\keyword{datasets}
