% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureSurvUnosC.R
\docType{data}
\name{MeasureSurvUnosC}
\alias{MeasureSurvUnosC}
\alias{mlr_measures_surv.unos_c}
\title{Uno's C-Index}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureSurv}.}
\usage{
MeasureSurvUnosC
}
\description{
Calls \code{\link[survAUC:UnoC]{survAUC::UnoC()}}.
}
\keyword{datasets}
