% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3survival-package}
\alias{mlr3survival}
\alias{mlr3survival-package}
\title{mlr3survival: Survival Analysis for mlr3}
\description{
Provides extensions for survival analysis for
    'mlr3'.  This includes a specialized task type, models, predictions,
    and measures.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mlr3survival.mlr-org.com}
  \item \url{https://github.com/mlr-org/mlr3survival}
  \item Report bugs at \url{https://github.com/mlr-org/mlsurvival/issues}
}

}
\author{
\strong{Maintainer}: Michel Lang \email{michellang@gmail.com} (0000-0001-9754-0393)

}
