% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_learners_surv_coxph.R
\docType{data}
\name{mlr_learners_surv.coxph}
\alias{mlr_learners_surv.coxph}
\alias{LearnerSurvCoxPH}
\title{Cox Proportional Hazard Learner}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{LearnerSurv}.}
\usage{
LearnerSurvCoxPH
}
\description{
A learner for a Cox PH model implemented in \link[survival:coxph]{survival::coxph}.
}
\keyword{datasets}
