% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_learners_surv_rpart.R
\docType{data}
\name{mlr_learners_surv.rpart}
\alias{mlr_learners_surv.rpart}
\alias{LearnerSurvRpart}
\title{Survival Tree Learner}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{LearnerSurv}.}
\usage{
LearnerSurvRpart
}
\description{
A learner for a regression tree implemented in \link[rpart:rpart]{rpart::rpart}.
}
\keyword{datasets}
