% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_learners_surv_ranger.R
\docType{data}
\name{mlr_learners_surv_ranger}
\alias{mlr_learners_surv_ranger}
\alias{LearnerSurvRanger}
\title{Survival Ranger Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{LearnerSurv}.}
\usage{
LearnerSurvRanger
}
\description{
A learner for a survival random forest implemented in \code{\link[ranger:ranger]{ranger::ranger()}}.
}
\keyword{datasets}
