% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_measures_surv_harrells_c.R
\docType{data}
\name{mlr_measures_surv.harrells_c}
\alias{mlr_measures_surv.harrells_c}
\alias{MeasureSurvHarrellsC}
\title{Harrell's C-Index}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureSurv}.}
\usage{
MeasureSurvHarrellsC
}
\description{
Calls \code{\link[Hmisc:rcorr.cens]{Hmisc::rcorr.cens()}}.
}
\keyword{datasets}
