% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskSurv_rats.R
\name{mlr_tasks_rats}
\alias{mlr_tasks_rats}
\alias{load_rats}
\title{Rats Survival Task}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{TaskSurv}.}
\description{
A survival task for the \link[survival:rats]{survival::rats} data set.
Columns have been converted to integer, column "sex" has been converted to a factor.
}
\section{Usage}{
\preformatted{mlr_tasks$get("rats")
}
}

