% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionSurv.R
\docType{data}
\name{PredictionSurv}
\alias{PredictionSurv}
\title{Prediction Object for Survival}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link[mlr3:Prediction]{mlr3::Prediction}.}
\description{
This object stores the predictions returned by a learner of class \link{LearnerSurv}.

The \code{task_type} is set to \code{"surv"}.
}
\section{Construction}{
\preformatted{p = PredictionSurv$new(task = NULL, risk = NULL)
}
\itemize{
\item \code{task} :: \link{TaskClassif}\cr
Task for which the predictions are made. Used to extract the row ids and the true
labels. Must be subsetted to test set.
\item \code{risk} :: \code{numeric()}\cr
Vector of risk scores. One element for each observation in the test set.
}

Note that it is allowed to initialize this object without any arguments in order
to allow to manually construct \link[mlr3:Prediction]{mlr3::Prediction} objects in a piecemeal fashion.
Required are "row_ids", "truth", and "predict_type". Depending on the value of
"predict_types", also "risk" must be set.
}

\section{Fields}{


\itemize{
\item \code{row_ids} :: (\code{integer()} | \code{character()})\cr
Vector of row ids for which predictions are stored.
\item \code{truth} :: \code{any}\cr
Vector of true labels.
\item \code{response} :: \code{any}\cr
Vector of predicted labels.
\item \code{task_type} :: \code{character(1)}\cr
Stores the type of the \link{Task}.
\item \code{predict_types} :: \code{character()}\cr
Vector of predict types this object stores.
}

}

\examples{
library(mlr3)
task = mlr_tasks$get("lung")
learner = mlr_learners$get("surv.rpart")
e = Experiment$new(task, learner)$train()$predict()
p = e$prediction
head(as.data.table(p))
}
\concept{Prediction}
\keyword{datasets}
