% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskSurvival.R
\name{TaskSurvival}
\alias{TaskSurvival}
\title{Survival Task}
\format{\link{R6Class} object inheriting from \link[mlr3:Task]{mlr3::Task}/\link[mlr3:TaskSupervised]{mlr3::TaskSupervised}.}
\description{
This task specializes \link[mlr3:Task]{mlr3::Task} and \link[mlr3:TaskSupervised]{mlr3::TaskSupervised} for right-censored survival problems.
}
\section{Usage}{

Inherits from [mlr3::Task]/[mlr3::TaskSupervised].\preformatted{# Construction
t = TaskSupervised$new(id, backend, time, status)


[mlr3::Task]: R:mlr3::Task
[mlr3::TaskSupervised]: R:mlr3::TaskSupervised
}
}

\section{Arguments}{

\itemize{
\item \code{time} (\code{numeric()}):
Time of event.
\item \code{status} (\code{numeric()}):
Event indicator. 0 means "alive" (no event), 1 means dead (event).
}
}

\examples{
data("lung", package = "survival")
b = as_data_backend(lung)
task = TaskSurvival$new("lung", backend = b, time = "time", status = "status")

task$target_names
task$feature_names
task$formula
}
