% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerSurv.R
\docType{data}
\name{LearnerSurv}
\alias{LearnerSurv}
\title{Survival Learner}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link[mlr3:Learner]{mlr3::Learner}.}
\description{
This Learner specializes \link[mlr3:Learner]{mlr3::Learner} for survival problems.
The slot \code{task_type} is set to \code{"surv"}.
Predefined learners can be found in the \link[mlr3:Dictionary]{mlr3::Dictionary} \link[mlr3:mlr_learners]{mlr3::mlr_learners}.
}
\section{Construction}{
\preformatted{l = LearnerSurv$new(id, param_set = ParamSet$new(), param_vals = list(), predict_types = character(),
     feature_types = character(), properties = character(), packages = character())
}

For a description of the arguments, see \link[mlr3:Learner]{mlr3::Learner}.
\code{task_type} is set to \code{"surv"}.
Currently only possible value for \code{predict_types} is "risk".
}

\section{Fields}{

See \link[mlr3:Learner]{mlr3::Learner}.
}

\section{Methods}{

See \link[mlr3:Learner]{mlr3::Learner}.
}

\examples{
library(mlr3)
ids = mlr_learners$keys("^surv")
lrns = mlr_learners$mget(ids)
names(lrns)

# get a specific learner from mlr_learners:
lrn = mlr_learners$get("surv.rpart")
print(lrn)
}
\concept{Learner}
\keyword{datasets}
