% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerSurvFeatureless.R
\docType{data}
\name{LearnerSurvFeatureless}
\alias{LearnerSurvFeatureless}
\alias{mlr_learners_surv.featureless}
\title{Featureless Survival Learner}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{LearnerSurv}.}
\usage{
LearnerSurvFeatureless
}
\description{
A simple \link{LearnerSurv} which only considers the survival time, ignoring all features.
Currently, only risk prediction is supported, where the learner just samples random
risk scores.
}
\keyword{datasets}
