% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureSurv.R
\name{MeasureSurv}
\alias{MeasureSurv}
\title{Survival Measure}
\format{\link{R6Class} object inheriting from \link{Measure}.}
\description{
This task specializes \link{Measure} for survival analysis problems.
}
\section{Usage}{

See \link{Measure}.
}

\section{Details}{

\code{$task_type} is \code{"surv"}.
}

