% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureSurvHarrellsC.R
\docType{data}
\name{MeasureSurvHarrellsC}
\alias{MeasureSurvHarrellsC}
\alias{mlr_measures_surv.harrells_c}
\title{Harrell's C-Index}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureSurv}.}
\usage{
MeasureSurvHarrellsC
}
\description{
Calculates Harrell's C, equivalent to the Fortran implementation in \CRANpkg{Hmisc}.
}
\keyword{datasets}
