% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskSurv_lung.R
\name{mlr_tasks_lung}
\alias{mlr_tasks_lung}
\alias{load_lung}
\title{Lung Cancer Survival Task}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{TaskSurv}.}
\description{
A survival task for the \link[survival:lung]{survival::lung} data set.
Columns have been converted to integer, column "sex" has been converted to a factor.
}
\section{Usage}{
\preformatted{mlr_tasks$get("lung")
}
}

