% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_tasks_surv_rats.R
\name{mlr_tasks_rats}
\alias{mlr_tasks_rats}
\alias{load_rats}
\title{Rats Survival Task}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{TaskSurv}.}
\usage{
load_rats()
}
\description{
A survival task for the \link[survival:rats]{survival::rats} data set.
}
\section{Usage}{
\preformatted{mlr_tasks$get("rats")
}
}

