% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionSurv.R
\name{PredictionSurv}
\alias{PredictionSurv}
\title{Prediction Object for Survival}
\format{\link{R6Class} object inheriting from \link{Prediction}.}
\description{
This object stores the predictions returned by a learner of class \link{LearnerSurv}.
}
\section{Usage}{

Inherits from \link{Prediction}\preformatted{# Construction
p = PredictionSurv$new(task, risk)

# Members
p$predict_types
p$risk
p$row_ids
p$truth

# S3 methods
as.data.table(p)
}
}

\section{Arguments}{

\itemize{
\item \code{task} (\link{Task}):
Task used for prediction. Used to extract \code{row_ids} and \code{truth}.
Set to \code{NULL} to skip all argument checks during initialization.
Slots \code{p$row_ids} and \code{p$truth} need to be set manually in this case
\item \code{risk} (\code{numeric()}): Vector of risk scores.
}
}

\section{Details}{

\itemize{
\item \code{$new()} initializes a new object of class \link{Prediction}.
\item \code{$predict_types} (\link{character}) stores the predict types available: currently only "risk" is supported.
\item \code{$risk} stores the predicted risk scores.
\item \code{row_ids} stores the row IDs.
\item \code{$truth} stores the true survival times as \code{\link[survival:Surv]{survival::Surv()}} object.
\item The prediction object can be transformed to a simple \code{\link[=data.table]{data.table()}}
with \link[data.table:as.data.table]{data.table::as.data.table}.
True survival time and status are separate columns, named "time" and "status".
}
}

\examples{
library(mlr3)
task = mlr_tasks$get("lung")
learner = mlr_learners$get("surv.rpart")
e = Experiment$new(task, learner)$train()$predict()
p = e$prediction
head(as.data.table(p))
}
\concept{Prediction}
